requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
websockets
psutil
pydantic
aiohttp